Add -Xverify:none to tests and docs
On newer versions of java bytecode is verified before it is run.  Since
this agent modifies core java bytecode it needs to bypass this
verification.  To turn the verification off pass -Xverify:none as an
argument when running the agent.  Fixed in unit tests and docs.

Details: https://www.ibm.com/support/pages/ibm-java-linux-howto-resolving-javalangverifyerror-jvmvrfy012-stack-shape-inconsistent